+++ src/DamYou/Views/PhotoPropertiesPanel.xaml.cs
@@ -1,0 +1,9 @@
+namespace DamYou.Views;
+
+public partial class PhotoPropertiesPanel : ContentView
+{
+    public PhotoPropertiesPanel()
+    {
+        InitializeComponent();
+    }
+}